Split ruststep::place_holder submodule
