+++ ruststep/src/lib.rs
@@ -15,6 +15,7 @@ pub mod ast;
 pub mod error;
 pub mod header;
 pub mod parser;
+pub mod place_holder;
 pub mod primitive;
 pub mod tables;
